feat: blocks migration
